+++ NEWS.md
@@ -15,3 +15,9 @@ Updated from package tdllicor, now runs on current versions of R.
 Resolve all R CHECK errors, warnings, notes.
 Updated all documentation using Roxygen.
 Modern package update complete.
+
+# isogasex 0.2.02
+
+6/9/2019
+Two versionf of input parameter file: yaml and xls.
+When copying decimal values from xls to yaml, copy full number from value bar (16 digits) instead of the 10 digits from the table cell.